+++ iroh-net/Cargo.toml
@@ -88,8 +88,6 @@ ntest = "*"
 pretty_assertions = "1.4"
 tokio = { version = "1", features = ["io-util", "sync", "rt", "net", "fs", "macros", "time", "test-util"] }
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
-strum = { version = "0.25.0", features = ["derive"]}
-ntest = "*"
 
 [build-dependencies]
 duct = "0.13.6"
